Delete Anagram/Nutrimatic sidebar tools from the menu.
These tools aren't any more convenient than just opening a new tab. Also, I removed them from the menu last year and just forgot to commit the change, and now I don't remember why I did that. Probably the feature was broken somehow and we didn't bother to fix it.

In any case, nobody has asked for this feature back (or, if they did ask, I lost the file with the feature requests in it) so let's just bury it.

This makes a bunch of code dead (including all of html/, I think), but we can delete that later, after Hunt.